Bedre navngiving og forenklinger
